segment crud
